(File) move download viewer logic to entities directory
